# services/gateway-core/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MeowShadow Lab Gateway Core API",
    "description": "Golang Fiber API Gateway, Authentication Hub, HTTP 206 Range Audio Streaming, Static Subtitles & Waveform Assets, and WebSocket Realtime Hub.",
    "version": "1.0.0",
    "contact": {
      "name": "MeowShadow Lab Engineering Team",
      "url": "https://github.com/tranviet-tv/MeowShadow_Lab"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Gateway"
    }
  ],
  "paths": {
...
    "/ws/progress": {
      "get": {
        "summary": "WebSocket Realtime Progress Hub",
        "tags": ["Realtime WebSockets"],
        "description": "WebSocket endpoint streaming live render progress percentages (15% -> 60% -> 90% -> 100%) and estimated remaining seconds.",
        "parameters": [
          { "name": "job_id", "in": "query", "description": "Filter progress events by Render Job UUID", "schema": { "type": "string" } },
          { "name": "lesson_id", "in": "query", "description": "Filter progress events by Lesson UUID", "schema": { "type": "string" } }
        ],
        "responses": {
          "101": { "description": "Switching Protocols to WebSocket" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}